+++ src/components/chat.tsx
@@ -1,23 +1,32 @@
-import { useState } from "react";
-import { Button } from "./ui/button";
-import { Textarea } from "./ui/textarea";
+import { useAtom } from "jotai";
+import { roleAtom } from "../store/role";
+import { useEffect, useState } from "react";
+import { sendChatAtom } from "../store/sendChat";
+import { chatLogAtom } from "../store/chatLog";
 import { getAnswer } from "../api/gpt";
-
-interface ChatEntry {
-  userMessage: string;
-  aiResponse: string | null;
-}
+import ChatInput from "./chatInput";
+import ChatLog from "./chatLog";
 
 export default function Chat() {
+  const [role] = useAtom(roleAtom);
+  const [isSend, setIsSend] = useAtom(sendChatAtom);
   const [textAreaValue, setTextAreaValue] = useState("");
-  const [chatLog, setChatLog] = useState<ChatEntry[]>([]);
+  const [, setChatLog] = useAtom(chatLogAtom);
 
   const handleChange = (event: React.ChangeEvent<HTMLTextAreaElement>) => {
     setTextAreaValue(event.target.value);
   };
 
-  const clickButton = async () => {
+  useEffect(() => {
+    setIsSend(false);
+  }, [setIsSend]);
+
+  const slicedRole = role[0]?.slice(0, 40) + "...";
+
+  const onClickSendButton = async () => {
     if (!textAreaValue.trim()) return;
+
+    setIsSend(true);
 
     setChatLog((prevChatLog) => [
       ...prevChatLog,
@@ -41,43 +50,36 @@ export default function Chat() {
   };
 
   return (
-    <div className="w-[85%] flex flex-col mt-10 p-6 space-y-10">
-      <div className="flex flex-col p-8 w-[85%] mx-auto my-0 bg-secondary rounded-xl space-y-7">
-        <div>
-          <p className="font-semibold text-white">🤖 AI 코치</p>
-          <p className="font-light text-white p-3">무엇을 도와드릴까요?</p>
-        </div>
-        {chatLog.map((entry, index) => (
-          <div key={index} className="flex flex-col space-y-4">
-            <div>
-              <p className="font-semibold text-white">🙂 YOU</p>
-              <p className="font-light text-white p-3">{entry.userMessage}</p>
-            </div>
-            <div>
-              <p className="font-semibold text-white">🤖 AI 코치</p>
-              <p className="font-light text-white p-3">
-                {entry.aiResponse === null
-                  ? "답변을 준비중입니다..."
-                  : entry.aiResponse}
+    <>
+      {isSend ? (
+        <ChatLog />
+      ) : (
+        <div className="w-[1116px] flex flex-col jusfity-center mx-auto my-auto items-center gap-7">
+          {role.length !== 0 ? (
+            <div className="flex flex-col items-center">
+              <div className="flex gap-3">
+                <p className="text-white text-2xl font-medium">{slicedRole} </p>
+                <p className="text-white text-2xl font-semibold">
+                  의 역할이 지정되어 있습니다.
+                </p>
+              </div>
+              <p className="text-white text-2xl font-semibold">
+                피드백을 요청할 코드를 입력해주세요.
               </p>
             </div>
-          </div>
-        ))}
-        <div className="flex items-center rounded-xl space-x-4 mt-10 bg-black">
-          <Textarea
-            value={textAreaValue}
-            onChange={handleChange}
-            placeholder="AI PRO에게 코드를 질문해보세요."
+          ) : (
+            <p className="text-white text-2xl font-semibold">
+              AI-PRO에 요청할 피드백과 코드를 입력해주세요.
+            </p>
+          )}
+          <ChatInput
+            textAreaValue={textAreaValue}
+            handleChange={handleChange}
+            onClickSendButton={onClickSendButton}
+            size="big"
           />
-          <Button
-            variant="default"
-            className="bg-black text-white h-[100%]"
-            onClick={clickButton}
-          >
-            질문하기
-          </Button>
         </div>
-      </div>
-    </div>
+      )}
+    </>
   );
 }
