+++ src/pages/Main.tsx
@@ -1,14 +1,12 @@
 import SideBar from "../components/sideBar";
-import Chat from "../components/chat";
 import Header from "../components/header";
-import FirstChat from "../components/firstChat";
+import Chat from "../components/chat";
 
 export default function Main() {
   return (
     <div className="flex flex-col bg-primary min-h-[100vh]">
       <Header />
-      <FirstChat />
-      {/*<Chat />*/}
+      <Chat />
       {/*<SideBar />*/}
     </div>
   );
